Fix packet counter variables in example NFs (#11)
@gkatsikas noticed that the packet counter variables in the example NFs were of type `int`.  When processing a large number of packets, an int was not large enough to store the count.

This is fixed by updating the variable type to `uint64_t` and updating the stats `printf`s.

Commit log:

* Fixed inconsistent packet counters in 4 examples

* Fixed format specifiers

* Three more cases are fixed

+++ examples/basic_monitor/monitor.c
@@ -112,7 +112,7 @@ static void
 do_stats_display(struct rte_mbuf* pkt) {
         const char clr[] = { 27, '[', '2', 'J', '\0' };
         const char topLeft[] = { 27, '[', '1', ';', '1', 'H', '\0' };
-        static int pkt_process = 0;
+        static uint64_t pkt_process = 0;
         struct ipv4_hdr* ip;
 
         pkt_process += print_delay;
@@ -125,7 +125,7 @@ do_stats_display(struct rte_mbuf* pkt) {
         printf("Port : %d\n", pkt->port);
         printf("Size : %d\n", pkt->pkt_len);
         printf("Hash : %u\n", pkt->hash.rss);
-	printf("N°   : %d\n", pkt_process);
+	printf("N°   : %"PRIu64"\n", pkt_process);
         printf("\n\n");
 
         ip = onvm_pkt_ipv4_hdr(pkt);
